Change return type to float32 in app.py

+++ app.py
@@ -105,7 +105,8 @@ def encode_texts(texts):
         counts = attention_mask.sum(dim=1).clamp(min=1)
         sentence_embs = summed / counts
         sentence_embs = nn.functional.normalize(sentence_embs, p=2, dim=1)
-    return sentence_embs.detach().cpu().numpy()
+
+    return sentence_embs.detach().cpu().to(torch.float32).numpy()
 
 
 SKILLS = {
